Remove sequential inquiry from BMax inquiry form
Because BMaX implementations should be implemented in a way which
requires as little knowledge as possible about the board its running
on so it can be used on many boards, by definition it cannot switch the
bus and thus cannot use sequential bus discovery. So the sequential
inquiry button was removed.